+++ tools/rosetta/lib/internal/service/network.go
@@ -1,0 +1,42 @@
+package service
+
+import (
+	"context"
+
+	"cosmossdk.io/tools/rosetta/lib/errors"
+	"github.com/coinbase/rosetta-sdk-go/types"
+)
+
+func (on OnlineNetwork) NetworkList(_ context.Context, _ *types.MetadataRequest) (*types.NetworkListResponse, *types.Error) {
+	return &types.NetworkListResponse{NetworkIdentifiers: []*types.NetworkIdentifier{on.network}}, nil
+}
+
+func (on OnlineNetwork) NetworkOptions(_ context.Context, _ *types.NetworkRequest) (*types.NetworkOptionsResponse, *types.Error) {
+	return on.networkOptions, nil
+}
+
+func (on OnlineNetwork) NetworkStatus(ctx context.Context, _ *types.NetworkRequest) (*types.NetworkStatusResponse, *types.Error) {
+	syncStatus, err := on.client.Status(ctx)
+	if err != nil {
+		return nil, errors.ToRosetta(err)
+	}
+
+	block, err := on.client.BlockByHeight(ctx, syncStatus.CurrentIndex)
+	if err != nil {
+		return nil, errors.ToRosetta(err)
+	}
+
+	peers, err := on.client.Peers(ctx)
+	if err != nil {
+		return nil, errors.ToRosetta(err)
+	}
+
+	return &types.NetworkStatusResponse{
+		CurrentBlockIdentifier: block.Block,
+		CurrentBlockTimestamp:  block.MillisecondTimestamp,
+		GenesisBlockIdentifier: on.genesisBlockIdentifier,
+		OldestBlockIdentifier:  nil,
+		SyncStatus:             syncStatus,
+		Peers:                  peers,
+	}, nil
+}
